testing/web-platform/tests/event-timing/event-timing-retrieve-firstInput.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: firstInput entry should be buffered even without observer</title>
<button id='button'>Generate a 'click' event</button>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>

<script src=resources/event-timing-support.js></script>

<script>
  async_test(function(t) {
    clickAndBlockMain('button').then(wait).then(t.step_func_done(() => {
      assert_equals(performance.getEntriesByType('firstInput').length, 1,
        "There should be a firstInput entry in the performance timeline");
      const entry = performance.getEntriesByType('firstInput')[0];
      assert_equals(entry.name, 'click');
      assert_equals(entry.entryType, 'firstInput');
      assert_greater_than(entry.duration, 50,
        "The first input was a long one.");
    }));
  },
  "Event Timing: check firstInput after onload, observer, click, click."
  );
</script>
</html>
